Move UniFFI types to own, feature-gated module
While we need quite a bit of dependencies and custom impls for UniFFI support, there is
no reason to pull them in for Rust-only builds. Here we move the
UniFFI-specific types and trait impls to a dedicated module and hide it
behind a new `uniffi` feature.
