+++ .github/workflows/build.yml
@@ -22,10 +22,16 @@ jobs:
           profile: minimal
       - name: Build on Rust ${{ matrix.toolchain }}
         run: cargo build --verbose --color always
+      - name: Build with UniFFI support on Rust ${{ matrix.toolchain }}
+        run: cargo build --features uniffi --verbose --color always
       - name: Check release build on Rust ${{ matrix.toolchain }}
         run: cargo check --release --verbose --color always
+      - name: Check release build with UniFFI support on Rust ${{ matrix.toolchain }}
+        run: cargo check --release --features uniffi --verbose --color always
+      - name: Test on Rust ${{ matrix.toolchain }}
+        run: cargo test
+      - name: Test with UniFFI support on Rust ${{ matrix.toolchain }}
+        run: cargo test --features uniffi
       - name: Check formatting on Rust ${{ matrix.toolchain }}
         if: matrix.check-fmt
         run: rustup component add rustfmt && cargo fmt --all -- --check
-      - name: Test on Rust ${{ matrix.toolchain }}
-        run: cargo test
