+++ Cargo.toml
@@ -27,7 +27,9 @@ opt-level = 'z'     # Optimize for size.
 lto = true          # Enable Link Time Optimization
 codegen-units = 1   # Reduce number of codegen units to increase optimizations.
 panic = 'abort'     # Abort on panic
-strip = true        # Strip symbols from binary*
+
+[features]
+default = []
 
 [dependencies]
 lightning = { version = "0.0.115", features = ["max_level_trace", "std"] }
@@ -67,7 +69,7 @@ serde_json = { version = "1.0" }
 tokio = { version = "1", default-features = false, features = [ "rt-multi-thread", "time", "sync" ] }
 esplora-client = { version = "0.4", default-features = false }
 libc = "0.2"
-uniffi = { version = "0.23.0", features = ["build"] }
+uniffi = { version = "0.23.0", features = ["build"], optional = true }
 
 [dev-dependencies]
 electrsd = { version = "0.22.0", features = ["legacy", "esplora_a33e97e1", "bitcoind_23_0"] }
@@ -76,8 +78,7 @@ proptest = "1.0.0"
 regex = "1.5.6"
 
 [build-dependencies]
-uniffi = { version = "0.23.0", features = ["build", "cli"] }
-
+uniffi = { version = "0.23.0", features = ["build", "cli"], optional = true }
 
 [profile.release]
 panic = "abort"
