+++ build.rs
@@ -1,3 +1,4 @@
 fn main() {
+	#[cfg(feature = "uniffi")]
 	uniffi::generate_scaffolding("bindings/ldk_node.udl").unwrap();
 }
