+++ scripts/uniffi_bindgen_generate_python.sh
@@ -2,6 +2,6 @@
 BINDINGS_DIR="./bindings/python"
 UNIFFI_BINDGEN_BIN="cargo +nightly run --features=uniffi/cli --bin uniffi-bindgen"
 
-cargo +nightly build --release || exit 1
+cargo +nightly build --release --features uniffi || exit 1
 $UNIFFI_BINDGEN_BIN generate bindings/ldk_node.udl --language python -o "$BINDINGS_DIR" || exit 1
 cp ./target/release/libldk_node.dylib "$BINDINGS_DIR"/libldk_node.dylib || exit 1
