+++ src/types.rs
@@ -1,8 +1,5 @@
-use crate::error::Error;
-use crate::hex_utils;
 use crate::logger::FilesystemLogger;
 use crate::wallet::{Wallet, WalletKeysManager};
-use crate::UniffiCustomTypeConverter;
 
 use lightning::chain::chainmonitor;
 use lightning::chain::keysinterface::InMemorySigner;
@@ -10,23 +7,17 @@ use lightning::ln::channelmanager::ChannelDetails as LdkChannelDetails;
 use lightning::ln::msgs::NetAddress as LdkNetAddress;
 use lightning::ln::msgs::RoutingMessageHandler;
 use lightning::ln::peer_handler::IgnoringMessageHandler;
-use lightning::ln::{PaymentHash, PaymentPreimage, PaymentSecret};
 use lightning::routing::gossip;
 use lightning::routing::router::DefaultRouter;
 use lightning::routing::scoring::ProbabilisticScorer;
 use lightning::util::ser::{Hostname, Readable, Writeable, Writer};
-use lightning_invoice::{Invoice, SignedRawInvoice};
 use lightning_net_tokio::SocketDescriptor;
 use lightning_transaction_sync::EsploraSyncClient;
 
-use bitcoin::hashes::sha256::Hash as Sha256;
-use bitcoin::hashes::Hash;
 use bitcoin::secp256k1::PublicKey;
-use bitcoin::{Address, OutPoint, Txid};
+use bitcoin::OutPoint;
 
-use bip39::Mnemonic;
-
-use std::convert::{TryFrom, TryInto};
+use std::convert::TryFrom;
 use std::fmt::Display;
 use std::net::{Ipv4Addr, Ipv6Addr, SocketAddr, SocketAddrV4, SocketAddrV6, ToSocketAddrs};
 use std::str::FromStr;
@@ -95,108 +86,6 @@ pub(crate) type OnionMessenger = lightning::onion_message::OnionMessenger<
 	IgnoringMessageHandler,
 >;
 
-impl UniffiCustomTypeConverter for PublicKey {
-	type Builtin = String;
-
-	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-		if let Ok(key) = PublicKey::from_str(&val) {
-			return Ok(key);
-		}
-
-		Err(Error::InvalidPublicKey.into())
-	}
-
-	fn from_custom(obj: Self) -> Self::Builtin {
-		obj.to_string()
-	}
-}
-
-impl UniffiCustomTypeConverter for Address {
-	type Builtin = String;
-
-	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-		if let Ok(addr) = Address::from_str(&val) {
-			return Ok(addr);
-		}
-
-		Err(Error::InvalidAddress.into())
-	}
-
-	fn from_custom(obj: Self) -> Self::Builtin {
-		obj.to_string()
-	}
-}
-
-impl UniffiCustomTypeConverter for Invoice {
-	type Builtin = String;
-
-	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-		if let Ok(signed) = val.parse::<SignedRawInvoice>() {
-			if let Ok(invoice) = Invoice::from_signed(signed) {
-				return Ok(invoice);
-			}
-		}
-
-		Err(Error::InvalidInvoice.into())
-	}
-
-	fn from_custom(obj: Self) -> Self::Builtin {
-		obj.to_string()
-	}
-}
-
-impl UniffiCustomTypeConverter for PaymentHash {
-	type Builtin = String;
-
-	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-		if let Ok(hash) = Sha256::from_str(&val) {
-			Ok(PaymentHash(hash.into_inner()))
-		} else {
-			Err(Error::InvalidPaymentHash.into())
-		}
-	}
-
-	fn from_custom(obj: Self) -> Self::Builtin {
-		Sha256::from_slice(&obj.0).unwrap().to_string()
-	}
-}
-
-impl UniffiCustomTypeConverter for PaymentPreimage {
-	type Builtin = String;
-
-	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-		if let Some(bytes_vec) = hex_utils::to_vec(&val) {
-			let bytes_res = bytes_vec.try_into();
-			if let Ok(bytes) = bytes_res {
-				return Ok(PaymentPreimage(bytes));
-			}
-		}
-		Err(Error::InvalidPaymentPreimage.into())
-	}
-
-	fn from_custom(obj: Self) -> Self::Builtin {
-		hex_utils::to_string(&obj.0)
-	}
-}
-
-impl UniffiCustomTypeConverter for PaymentSecret {
-	type Builtin = String;
-
-	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-		if let Some(bytes_vec) = hex_utils::to_vec(&val) {
-			let bytes_res = bytes_vec.try_into();
-			if let Ok(bytes) = bytes_res {
-				return Ok(PaymentSecret(bytes));
-			}
-		}
-		Err(Error::InvalidPaymentSecret.into())
-	}
-
-	fn from_custom(obj: Self) -> Self::Builtin {
-		hex_utils::to_string(&obj.0)
-	}
-}
-
 /// The global identifier of a channel.
 ///
 /// Note that this will start out to be a temporary ID until channel funding negotiation is
@@ -204,25 +93,6 @@ impl UniffiCustomTypeConverter for PaymentSecret {
 /// funding transaction.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 pub struct ChannelId(pub [u8; 32]);
-
-impl UniffiCustomTypeConverter for ChannelId {
-	type Builtin = String;
-
-	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-		if let Some(hex_vec) = hex_utils::to_vec(&val) {
-			if hex_vec.len() == 32 {
-				let mut channel_id = [0u8; 32];
-				channel_id.copy_from_slice(&hex_vec[..]);
-				return Ok(Self(channel_id));
-			}
-		}
-		Err(Error::InvalidChannelId.into())
-	}
-
-	fn from_custom(obj: Self) -> Self::Builtin {
-		hex_utils::to_string(&obj.0)
-	}
-}
 
 impl Writeable for ChannelId {
 	fn write<W: Writer>(&self, writer: &mut W) -> Result<(), lightning::io::Error> {
@@ -244,18 +114,6 @@ impl Readable for ChannelId {
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 pub struct UserChannelId(pub u128);
 
-impl UniffiCustomTypeConverter for UserChannelId {
-	type Builtin = String;
-
-	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-		Ok(UserChannelId(u128::from_str(&val).map_err(|_| Error::InvalidChannelId)?))
-	}
-
-	fn from_custom(obj: Self) -> Self::Builtin {
-		obj.0.to_string()
-	}
-}
-
 impl Writeable for UserChannelId {
 	fn write<W: Writer>(&self, writer: &mut W) -> Result<(), lightning::io::Error> {
 		Ok(self.0.write(writer)?)
@@ -267,17 +125,6 @@ impl Readable for UserChannelId {
 		reader: &mut R,
 	) -> Result<Self, lightning::ln::msgs::DecodeError> {
 		Ok(Self(Readable::read(reader)?))
-	}
-}
-
-impl UniffiCustomTypeConverter for Txid {
-	type Builtin = String;
-	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-		Ok(Txid::from_str(&val)?)
-	}
-
-	fn from_custom(obj: Self) -> Self::Builtin {
-		obj.to_string()
 	}
 }
 
@@ -422,28 +269,6 @@ impl Display for NetAddress {
 				ed25519_pubkey, checksum, version, port
 			),
 		}
-	}
-}
-
-impl UniffiCustomTypeConverter for Mnemonic {
-	type Builtin = String;
-	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-		Ok(Mnemonic::from_str(&val).map_err(|_| Error::InvalidSecretKey)?)
-	}
-
-	fn from_custom(obj: Self) -> Self::Builtin {
-		obj.to_string()
-	}
-}
-
-impl UniffiCustomTypeConverter for NetAddress {
-	type Builtin = String;
-	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-		Ok(NetAddress::from_str(&val).map_err(|_| Error::InvalidNetAddress)?)
-	}
-
-	fn from_custom(obj: Self) -> Self::Builtin {
-		obj.to_string()
 	}
 }
