+++ src/uniffi_types.rs
@@ -1,0 +1,188 @@
+use crate::UniffiCustomTypeConverter;
+
+use crate::error::Error;
+use crate::hex_utils;
+use crate::io::FilesystemStore;
+use crate::{ChannelId, NetAddress, Node, UserChannelId};
+
+use bitcoin::hashes::sha256::Hash as Sha256;
+use bitcoin::hashes::Hash;
+use bitcoin::secp256k1::PublicKey;
+use bitcoin::{Address, Txid};
+use lightning::ln::{PaymentHash, PaymentPreimage, PaymentSecret};
+use lightning_invoice::{Invoice, SignedRawInvoice};
+
+use bip39::Mnemonic;
+
+use std::convert::TryInto;
+use std::str::FromStr;
+
+/// This type alias is required as Uniffi doesn't support generics, i.e., we can only expose the
+/// concretized types via this aliasing hack.
+pub type LDKNode = Node<FilesystemStore>;
+
+impl UniffiCustomTypeConverter for PublicKey {
+	type Builtin = String;
+
+	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
+		if let Ok(key) = PublicKey::from_str(&val) {
+			return Ok(key);
+		}
+
+		Err(Error::InvalidPublicKey.into())
+	}
+
+	fn from_custom(obj: Self) -> Self::Builtin {
+		obj.to_string()
+	}
+}
+
+impl UniffiCustomTypeConverter for Address {
+	type Builtin = String;
+
+	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
+		if let Ok(addr) = Address::from_str(&val) {
+			return Ok(addr);
+		}
+
+		Err(Error::InvalidAddress.into())
+	}
+
+	fn from_custom(obj: Self) -> Self::Builtin {
+		obj.to_string()
+	}
+}
+
+impl UniffiCustomTypeConverter for Invoice {
+	type Builtin = String;
+
+	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
+		if let Ok(signed) = val.parse::<SignedRawInvoice>() {
+			if let Ok(invoice) = Invoice::from_signed(signed) {
+				return Ok(invoice);
+			}
+		}
+
+		Err(Error::InvalidInvoice.into())
+	}
+
+	fn from_custom(obj: Self) -> Self::Builtin {
+		obj.to_string()
+	}
+}
+
+impl UniffiCustomTypeConverter for PaymentHash {
+	type Builtin = String;
+
+	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
+		if let Ok(hash) = Sha256::from_str(&val) {
+			Ok(PaymentHash(hash.into_inner()))
+		} else {
+			Err(Error::InvalidPaymentHash.into())
+		}
+	}
+
+	fn from_custom(obj: Self) -> Self::Builtin {
+		Sha256::from_slice(&obj.0).unwrap().to_string()
+	}
+}
+
+impl UniffiCustomTypeConverter for PaymentPreimage {
+	type Builtin = String;
+
+	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
+		if let Some(bytes_vec) = hex_utils::to_vec(&val) {
+			let bytes_res = bytes_vec.try_into();
+			if let Ok(bytes) = bytes_res {
+				return Ok(PaymentPreimage(bytes));
+			}
+		}
+		Err(Error::InvalidPaymentPreimage.into())
+	}
+
+	fn from_custom(obj: Self) -> Self::Builtin {
+		hex_utils::to_string(&obj.0)
+	}
+}
+
+impl UniffiCustomTypeConverter for PaymentSecret {
+	type Builtin = String;
+
+	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
+		if let Some(bytes_vec) = hex_utils::to_vec(&val) {
+			let bytes_res = bytes_vec.try_into();
+			if let Ok(bytes) = bytes_res {
+				return Ok(PaymentSecret(bytes));
+			}
+		}
+		Err(Error::InvalidPaymentSecret.into())
+	}
+
+	fn from_custom(obj: Self) -> Self::Builtin {
+		hex_utils::to_string(&obj.0)
+	}
+}
+
+impl UniffiCustomTypeConverter for ChannelId {
+	type Builtin = String;
+
+	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
+		if let Some(hex_vec) = hex_utils::to_vec(&val) {
+			if hex_vec.len() == 32 {
+				let mut channel_id = [0u8; 32];
+				channel_id.copy_from_slice(&hex_vec[..]);
+				return Ok(Self(channel_id));
+			}
+		}
+		Err(Error::InvalidChannelId.into())
+	}
+
+	fn from_custom(obj: Self) -> Self::Builtin {
+		hex_utils::to_string(&obj.0)
+	}
+}
+
+impl UniffiCustomTypeConverter for UserChannelId {
+	type Builtin = String;
+
+	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
+		Ok(UserChannelId(u128::from_str(&val).map_err(|_| Error::InvalidChannelId)?))
+	}
+
+	fn from_custom(obj: Self) -> Self::Builtin {
+		obj.0.to_string()
+	}
+}
+
+impl UniffiCustomTypeConverter for Txid {
+	type Builtin = String;
+	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
+		Ok(Txid::from_str(&val)?)
+	}
+
+	fn from_custom(obj: Self) -> Self::Builtin {
+		obj.to_string()
+	}
+}
+
+impl UniffiCustomTypeConverter for Mnemonic {
+	type Builtin = String;
+	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
+		Ok(Mnemonic::from_str(&val).map_err(|_| Error::InvalidSecretKey)?)
+	}
+
+	fn from_custom(obj: Self) -> Self::Builtin {
+		obj.to_string()
+	}
+}
+
+impl UniffiCustomTypeConverter for NetAddress {
+	type Builtin = String;
+	fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
+		Ok(NetAddress::from_str(&val).map_err(|_| Error::InvalidNetAddress)?)
+	}
+
+	fn from_custom(obj: Self) -> Self::Builtin {
+		obj.to_string()
+	}
+}
