+++ uniffi-bindgen.rs
@@ -1,3 +1,4 @@
 fn main() {
+	#[cfg(feature = "uniffi")]
 	uniffi::uniffi_bindgen_main()
 }
